include-src structure recreated

+++ include/particle.h
@@ -10,15 +10,15 @@ public:
     /// @brief Adjust the position of the particle.
     /// @param change The change in position.
     /// @param dimension The dimension to change.
-    void adjustPosition(double change, unsigned int dimension);
+    void adjustPosition(double change, size_t dimension);
     /// @brief Get the position of the particle.
     /// @return The position of the particle.
     std::vector<double> getPosition() { return m_position; }
     /// @brief Get the number of dimensions.
     /// @return Number of dimensions 
-    unsigned int getNumberOfDimensions() { return m_numberOfDimensions; }
+    size_t getNumberOfDimensions() { return m_numberOfDimensions; }
 
 private:
-    unsigned int m_numberOfDimensions = 0;
+    size_t m_numberOfDimensions = 0;
     std::vector<double> m_position = std::vector<double>();
 };
