+++ include/sampler.h
@@ -5,10 +5,10 @@ class Sampler
 {
 public:
     Sampler(
-        unsigned int numberOfParticles,
-        unsigned int numberOfDimensions,
+        size_t numberOfParticles,
+        size_t numberOfDimensions,
         double stepLength,
-        unsigned int numberOfMetropolisSteps);
+        size_t numberOfMetropolisSteps);
 
     void sample(bool acceptedStep, class System *system);
     void printOutputToTerminal(class System &system);
@@ -16,11 +16,11 @@ public:
     double getEnergy() { return m_energy; }
 
 private:
-    unsigned int m_stepNumber = 0;
-    unsigned int m_numberOfMetropolisSteps = 0;
-    unsigned int m_numberOfParticles = 0;
-    unsigned int m_numberOfDimensions = 0;
-    unsigned int m_numberOfAcceptedSteps = 0;
+    size_t m_stepNumber = 0;
+    size_t m_numberOfMetropolisSteps = 0;
+    size_t m_numberOfParticles = 0;
+    size_t m_numberOfDimensions = 0;
+    size_t m_numberOfAcceptedSteps = 0;
     double m_energy = 0;
     double m_cumulativeEnergy = 0;
     double m_stepLength = 0;
